requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.5.0
pygame>=2.1.0
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
pyserial>=3.5
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0
